Add a "framework" folder to prevent package name clashes
