+++ example/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 group = "com.jeantuffier"
-version = "2.1.0-alpha1"
+version = "2.1.0-alpha2"
 
 repositories {
     google()
