+++ example/src/commonMain/kotlin/com/jeantuffier/statemachine/School.kt
@@ -3,6 +3,9 @@ package com.jeantuffier.statemachine
 import com.jeantuffier.statemachine.annotation.CrossStateProperty
 import com.jeantuffier.statemachine.annotation.ViewEventsBuilder
 import com.jeantuffier.statemachine.annotation.ViewState
+import com.jeantuffier.statemachine.framework.AsyncData
+import com.jeantuffier.statemachine.framework.StateMachine
+import com.jeantuffier.statemachine.framework.StateMachineBuilder
 
 @ViewState
 data class SchoolViewState(
