+++ example/src/commonMain/kotlin/com/jeantuffier/statemachine/SchoolStaff.kt
@@ -3,6 +3,7 @@ package com.jeantuffier.statemachine
 import com.jeantuffier.statemachine.annotation.CrossStateProperty
 import com.jeantuffier.statemachine.annotation.ViewEventsBuilder
 import com.jeantuffier.statemachine.annotation.ViewState
+import com.jeantuffier.statemachine.framework.*
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.update
